requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-cleanup==5.2.0
django-heroku==0.3.1
django-stdimage==5.3.0
gunicorn==20.1.0
idna==2.10
Pillow==8.2.0
psycopg2==2.9.1
python-dateutil==1.5
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.2.0
